Move associateTileToBackground method into picture
