+++ packages/core/src/helper.js
@@ -1,16 +1,3 @@
-export const associateTileToBackground = grid => {
-    const tileSize = grid.length > 0 ? 100 / grid.length : 0; // 100%
-
-    let tileToBg = {};
-
-    grid.forEach((row, rowKey) => {
-        row.forEach((value, colKey) => {
-            tileToBg[value] = `${tileSize * colKey}% ${tileSize * rowKey}%`;
-        });
-    });
-    return tileToBg;
-};
-
 export const choiceInArray = (list, rand = Math.random()) => {
     if (!Array.isArray(list)) {
         throw Error('The list should be an array');
